Began added the encode function.
