+++ README.md
@@ -11,3 +11,7 @@ Licensed under either of:
 - [Apache License, Version 2.0](http://www.apache.org/licenses/LICENSE-2.0)
 - [MIT license](http://opensource.org/licenses/MIT)
 at your option.
+
+
+## TODO list:
+- Rename EncodedRle to COCO_RLE ?
