+++ rpycocotools/python/rpycocotools/__init__.py
@@ -1,2 +1,2 @@
 from rpycocotools.rpycocotools import *
-from python.rpycocotools.mask import decode
+import python.rpycocotools.mask as mask
