+++ rpycocotools/python/rpycocotools/mask/__init__.py
@@ -1,0 +1,1 @@
+from .mask import decode, encode
